% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{Title}
\usage{
mc(df.split, iteration, nc, sub, var1, var2, var3)
}
\arguments{
\item{df.split}{被分半的list}

\item{iteration}{迭代次数}

\item{nc}{使用的CPU线程数}

\item{sub}{被试是哪一列}

\item{var1}{第一个自变量}

\item{var2}{第二个自变量}

\item{var3}{第三个自变量}
}
\value{
返回蒙特卡洛分半的结果
}
\description{
Title
}
